requirements.txt
Django>=5.2.8
djangorestframework>=3.14
django-cors-headers>=3.14
gunicorn>=20.1.0
psycopg2-binary>=2.9.6
transformers>=4.44.0
sentencepiece>=0.1.99
numpy<2
torch==2.2.0